Bugfix: when a way is modified by its nodes:  * Nodes are moved or retagged  * No noded added, no node removed  * No tags changed
The augmented diff will contain:
  <action type="modify" geometry="changed" members="changed">
    <old>OLD WAY</old><new>OLD WAY</new>
  </action>

In that case, we must not consider the "OLD WAY" (and its associated chnageset).
So in that case, we look in the nodes of the way the ones that have been changed, and
we consider the way to belong to this changeset.

This fixes random apparitions of very old chnagesets in the watches